refactor(kernel-utils): centralize guard payload helpers and move asyncifyMethodGuards to guard.ts
